+++ ports/esp32/mphalport.h
@@ -61,8 +61,9 @@ __attribute__((always_inline)) static inline uint32_t mp_hal_ticks_cpu(void) {
 }
 
 #define mp_hal_ticks_cpu_enable(void)
-#define mp_hal_ticks_cpu_start mp_hal_ticks_cpu
-#define MP_HAL_BITSTREAM_NS_OVERHEAD  (6)
+#define mp_hal_ticks_bitstream mp_hal_ticks_cpu
+#define mp_hal_ticks_bitstream_start mp_hal_ticks_cpu
+#define MP_HAL_BITSTREAM_NS_OVERHEAD  (7)
 
 static inline mp_uint_t mp_hal_get_cpu_freq(void) {
     return ets_get_cpu_frequency() * 1000000;
