+++ ports/esp8266/esp_mphal.h
@@ -62,7 +62,8 @@ __attribute__((always_inline)) static inline uint32_t mp_hal_ticks_cpu(void) {
 }
 
 #define mp_hal_ticks_cpu_enable(void)
-#define mp_hal_ticks_cpu_start mp_hal_ticks_cpu
+#define mp_hal_ticks_bitstream_start mp_hal_ticks_cpu
+#define mp_hal_ticks_bitstream mp_hal_ticks_cpu
 #define MP_HAL_BITSTREAM_NS_OVERHEAD  (5)
 
 void mp_hal_delay_us(uint32_t);
@@ -111,11 +112,11 @@ void mp_hal_pin_open_drain(mp_hal_pin_obj_t pin);
 #define mp_hal_pin_write(p, v) pin_set((p), (v))
 
 static inline void mp_hal_pin_low(mp_hal_pin_obj_t pin) {
-        GPIO_REG_WRITE(GPIO_OUT_W1TC_ADDRESS, 1 << pin);
+    GPIO_REG_WRITE(GPIO_OUT_W1TC_ADDRESS, 1 << pin);
 }
 
 static inline void mp_hal_pin_high(mp_hal_pin_obj_t pin) {
-        GPIO_REG_WRITE(GPIO_OUT_W1TS_ADDRESS, 1 << pin);
+    GPIO_REG_WRITE(GPIO_OUT_W1TS_ADDRESS, 1 << pin);
 }
 void *ets_get_esf_buf_ctlblk(void);
 int ets_esf_free_bufs(int idx);
