+++ ports/rp2/mphalport.h
@@ -73,11 +73,14 @@ static inline mp_uint_t mp_hal_ticks_cpu_start(void) {
     return 0;
 }
 
+#define mp_hal_ticks_bitstream mp_hal_ticks_cpu
+#define mp_hal_ticks_bitstream_start mp_hal_ticks_cpu_start
+
 static inline mp_uint_t mp_hal_get_cpu_freq(void) {
     return clock_get_hz(clk_sys);
 }
 
-#define MP_HAL_BITSTREAM_NS_OVERHEAD  (9)
+#define MP_HAL_BITSTREAM_NS_OVERHEAD  (8)
 
 // C-level pin HAL
